backend/services/ai/vision.py
"""
AI Vision service for food recognition
"""
from typing import List, Dict, Any
import logging
from io import BytesIO
from PIL import Image

logger = logging.getLogger(__name__)

class FoodVisionService:
    """Food recognition service using computer vision"""
    
    def __init__(self):
        self.model_loaded = False
        # TODO: Initialize actual ML model here
        
    async def analyze_food_image(self, image_bytes: bytes) -> Dict[str, Any]:
        """
        Analyze food image and return detected items with nutritional estimates
        
        Args:
            image_bytes: Image data as bytes
            
        Returns:
            Dict containing detected foods with confidence scores and portion estimates
        """
        try:
            # TODO: Replace with actual AI model inference
            # This is a placeholder implementation
            
            # Validate image
            image = Image.open(BytesIO(image_bytes))
            if image.mode != 'RGB':
                image = image.convert('RGB')
            
            # Mock detection results
            mock_results = {
                "items": [
                    {
                        "name": "arroz blanco",
                        "confidence": 0.85,
                        "portion_guess_grams": 150.0,
                        "bounding_box": [0.2, 0.3, 0.6, 0.7],
                        "nutritional_estimate": {
                            "calories": 195,
                            "protein_g": 4.0,
                            "carbs_g": 40.0,
                            "fat_g": 0.5
                        }
                    },
                    {
                        "name": "pollo a la plancha",
                        "confidence": 0.92,
                        "portion_guess_grams": 120.0,
                        "bounding_box": [0.1, 0.1, 0.5, 0.5],
                        "nutritional_estimate": {
                            "calories": 198,
                            "protein_g": 37.0,
                            "carbs_g": 0.0,
                            "fat_g": 4.3
                        }
                    }
                ],
                "confidence_score": 0.88,
                "processing_time_ms": 245,
                "model_version": "v1.0.0-mock"
            }
            
            logger.info(f"Food vision analysis completed: {len(mock_results['items'])} items detected")
            return mock_results
            
        except Exception as e:
            logger.error(f"Error in food vision analysis: {e}")
            raise ValueError(f"Failed to analyze image: {str(e)}")
    
    async def classify_food_item(self, food_name: str, image_bytes: bytes) -> Dict[str, Any]:
        """
        Classify a specific food item and estimate portion size
        
        Args:
            food_name: Name of the food to classify
            image_bytes: Image data as bytes
            
        Returns:
            Classification result with portion estimation
        """
        # TODO: Implement specific food classification
        return {
            "food_name": food_name,
            "classification_confidence": 0.85,
            "estimated_portion_grams": 100.0,
            "portion_confidence": 0.70,
            "suggestions": [
                "Consider using a food scale for accurate measurement",
                "Portion appears to be standard serving size"
            ]
        }
    
    async def get_portion_estimation(self, image_bytes: bytes, reference_objects: List[str] = None) -> Dict[str, Any]:
        """
        Estimate portion sizes using reference objects in the image
        
        Args:
            image_bytes: Image data as bytes
            reference_objects: List of reference objects (e.g., "coin", "hand", "plate")
            
        Returns:
            Portion estimation results
        """
        # TODO: Implement portion estimation logic
        return {
            "reference_objects_detected": reference_objects or ["plate"],
            "scale_factor": 1.2,
            "portion_estimates": {
                "small": "50-75g",
                "medium": "100-125g", 
                "large": "150-200g"
            },
            "confidence": 0.75
        }

# Service instance
food_vision_service = FoodVisionService()

# Convenience functions
async def analyze_food_image(image_bytes: bytes) -> Dict[str, Any]:
    """Analyze food image and return detected items"""
    return await food_vision_service.analyze_food_image(image_bytes)

async def classify_food_item(food_name: str, image_bytes: bytes) -> Dict[str, Any]:
    """Classify specific food item"""
    return await food_vision_service.classify_food_item(food_name, image_bytes)

async def estimate_portions(image_bytes: bytes, reference_objects: List[str] = None) -> Dict[str, Any]:
    """Estimate portion sizes"""
    return await food_vision_service.get_portion_estimation(image_bytes, reference_objects)