backend/requirements.txt
# Core API
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
pydantic-settings==2.1.0

# Database
sqlmodel==0.0.14
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Auth & Security
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
argon2-cffi==23.1.0
fastapi-limiter==0.1.6
PyJWT==2.8.0
bcrypt==4.1.2
python-dotenv==1.0.0

# Async & Background Tasks
celery==5.3.4
redis==5.0.1

# External Services
google-cloud-storage==2.12.0
sendgrid==6.11.0

# AI/ML Placeholders
pillow==10.1.0
numpy==1.25.2

# Monitoring & Observability
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
structlog==23.2.0

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
black==23.11.0
ruff==0.1.7

# Utils
emails==0.6.0